chore: Align packages and clean up module structure (#1)
