--- eventus-core/bus/EventBus.java
@@ -1,6 +1,0 @@
-import java.beans.EventHandler;
-
-public interface EventBus {
-    void publish(Event event);
-    void subscribe(String eventType, EventHandler<? extends Event> handler);
-}
