--- eventus-core/dlq/DeadLetterPublisher.java
@@ -1,3 +1,0 @@
-public interface DeadLetterPublisher {
-    void publish(Event event, Exception cause);
-}
