--- eventus-core/handler/EventHandler.java
@@ -1,3 +1,0 @@
-public interface EventHandler<E extends Event> {
-    void handle(E event) throws Exception;
-}
