+++ eventus-core/src/main/java/io/github/oleandrorodrigues/eventus/dlq/DeadLetterPublisher.java
@@ -1,0 +1,7 @@
+package io.github.oleandrorodrigues.eventus.dlq;
+
+import io.github.oleandrorodrigues.eventus.event.Event;
+
+public interface DeadLetterPublisher {
+    void publish(Event event, Exception cause);
+}
