+++ eventus-core/src/main/java/io/github/oleandrorodrigues/eventus/event/Event.java
@@ -1,3 +1,5 @@
+package io.github.oleandrorodrigues.eventus.event;
+
 import java.time.Instant;
 import java.util.Map;
