+++ eventus-core/src/main/java/io/github/oleandrorodrigues/eventus/exception/NonRetryableException.java
@@ -1,1 +1,3 @@
+package io.github.oleandrorodrigues.eventus.exception;
+
 public class NonRetryableException extends RuntimeException {}
