+++ eventus-core/src/main/java/io/github/oleandrorodrigues/eventus/exception/RetryableException.java
@@ -1,1 +1,3 @@
+package io.github.oleandrorodrigues.eventus.exception;
+
 public class RetryableException extends RuntimeException {}
