+++ eventus-core/src/main/java/io/github/oleandrorodrigues/eventus/handler/EventHandler.java
@@ -1,0 +1,7 @@
+package io.github.oleandrorodrigues.eventus.handler;
+
+import io.github.oleandrorodrigues.eventus.event.Event;
+
+public interface EventHandler<E extends Event> {
+    void handle(E event) throws Exception;
+}
