+++ eventus-core/src/main/java/io/github/oleandrorodrigues/eventus/idempotency/IdempotencyStore.java
@@ -1,3 +1,5 @@
+package io.github.oleandrorodrigues.eventus.idempotency;
+
 public interface IdempotencyStore {
     boolean isProcessed(String eventId, String handlerName);
     void markProcessed(String eventId, String handlerName);
