--- eventus-examples/src/main/java/Main.java
@@ -1,7 +1,0 @@
-package io.github.oleandrorodrigues.eventus.examples;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Eventus examples running");
-    }
-}
